requirements.txt
torchinfo==1.8.0
torchview==0.2.6
accelerate==0.21.0
transformers==4.31.0
datasets==2.14.4
diffusers==0.21.4
seaborn==0.12.2
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.4
wandb==0.15.8
scikit-learn==1.3.0
tqdm==4.66.1
pyyaml==6.0.1
rich==13.5.2
huggingface_hub==0.16.4
pytest==7.4.0
onnx==1.14.0
onnxruntime==1.15.1
graphviz==0.20.1